Wrote readme with very basic ideas
I want to get this setup for stardrain, but before i get deep into the nitty gritty of functionality, I asked grok and cursor to give me a general idea of ways to get started. still need others to help me build this properly. ways to get verification of minigame achievements in sega roms embedded in some rooms would be interesting, but not necessary early on.